Refactor code to allow different paper sizes
This change introduces the concept of a Page that can have margins.
After this change it is now possible to have different layouts and
paper sizes, which can be added in later commits.

Now to add a new layout, you just need to pass in a layoutFunc
function to the renderPages function, that knows how each page
should lay out hte individual frames.

For now only 6x4 is supported, but letter support (8.5x11) will be
added in a later commit.