Add support for margins option
The different layout options all have default settings for their
margins that make sure they are the desired dimensions, but you may
need to change these values depending on your printer.

For example, my printer when selecting "full page, borderless"
actually scales up the original image so it prints outside the bounds
of the paper. This has the effect that if you are doing a 6x4 with 3
frames that are suppose to be 2" tall, the top and bottom frame are
clipped, so the middle frame is taller.

In this case I had to play around with my printer and margins until I
found margin values that allowed me to print all of the frame content
 without any clipping.