Add new video effects
This commit adds the oil, cartoon, edge and pencil effects, which can
be enabled by using the effect command line option.

Note that the quality of the effect on the output is very dependant
on the video, for example the cartoon effect may look good on one
video but not good on another, you have to play around and see if it
looks good for the video you are processing.